Refactor: 도미노 키보드 제어 로직 useDominoKeyboardControls 훅으로 이전 및 통합
- 기존 useDominoControls 내부에 있던 키 이벤트 핸들러 (삭제, 투명도 토글, Undo 등)를 useDominoKeyboardControls 훅으로 모두 이전
- keyMap 구조를 통해 키 입력 로직을 통합 및 관리
- 최신 상태를 반영하기 위해 useDominoStore.getState() 활용
- 불필요해진 useDominoControls 파일 제거

[Refactor] 도미노 키보드 이벤트 통합 및 회전 기능 추가 #50
